+++ example/lib/import_resolution_types.dart
@@ -1,0 +1,13 @@
+extension type OrgId._(String value) {
+  factory OrgId.fromValue(String value) => OrgId._(value);
+  String toJson() => value;
+  String call() => value;
+}
+
+extension type CustomerId._(String value) {
+  factory CustomerId.fromValue(String value) => CustomerId._(value);
+  String toJson() => value;
+  String call() => value;
+}
+
+enum PaymentChannel { cash, card, bankTransfer }