+++ example/lib/user_profile.widget.dart
@@ -75,8 +75,10 @@ class _AsyncUserProfileParamsInheritedWidget extends InheritedWidget {
   final int id;
 
   static _AsyncUserProfileParamsInheritedWidget? maybeOf(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<
-        _AsyncUserProfileParamsInheritedWidget>();
+    return context
+        .dependOnInheritedWidgetOfExactType<
+          _AsyncUserProfileParamsInheritedWidget
+        >();
   }
 
   static _AsyncUserProfileParamsInheritedWidget of(BuildContext context) {
@@ -121,11 +123,13 @@ class AsyncUserProfileScope extends ConsumerWidget {
     BuildContext context,
     AsyncValue<UserProfileState> asyncValue,
     Widget child,
-  )? builder;
+  )?
+  builder;
   final void Function(
     AsyncValue<UserProfileState>? previous,
     AsyncValue<UserProfileState> next,
-  )? onStateChanged;
+  )?
+  onStateChanged;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -233,11 +237,13 @@ class AsyncUserProfileWidget extends ConsumerWidget {
     BuildContext context,
     AsyncUserProfileProxyWidgetRef ref,
     AsyncValue<UserProfileState> state,
-  ) builder;
+  )
+  builder;
   final void Function(
     AsyncValue<UserProfileState>? previous,
     AsyncValue<UserProfileState> next,
-  )? onStateChanged;
+  )?
+  onStateChanged;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -287,7 +293,8 @@ class AsyncUserProfileSelect<Selected> extends ConsumerWidget {
     BuildContext context,
     AsyncUserProfileProxyWidgetRef ref,
     Selected value,
-  ) builder;
+  )
+  builder;
   final Widget? loading;
   final Widget Function(Object error, StackTrace stackTrace)? error;
   final void Function(Selected? previous, Selected next)? onStateChanged;
@@ -350,7 +357,8 @@ class AsyncUserProfileParamsBuilder extends ConsumerWidget {
     BuildContext context,
     AsyncUserProfileProxyWidgetRef ref,
     int params,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -405,7 +413,8 @@ class AsyncUserProfileNameField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     AsyncUserProfileNameProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
   final Widget? loading;
   final Widget Function(Object error, StackTrace stackTrace)? error;
 
@@ -487,7 +496,8 @@ class AsyncUserProfileEmailField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     AsyncUserProfileEmailProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
   final Widget? loading;
   final Widget Function(Object error, StackTrace stackTrace)? error;
 
@@ -570,7 +580,8 @@ class AsyncUserProfileAgeField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     AsyncUserProfileAgeProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
   final Widget? loading;
   final Widget Function(Object error, StackTrace stackTrace)? error;
 
@@ -651,7 +662,8 @@ class AsyncUserProfileBioField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     AsyncUserProfileBioProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
   final Widget? loading;
   final Widget Function(Object error, StackTrace stackTrace)? error;
 
@@ -732,7 +744,8 @@ class AsyncUserProfileBio2Field extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     AsyncUserProfileBio2ProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
   final Widget? loading;
   final Widget Function(Object error, StackTrace stackTrace)? error;
 
@@ -813,8 +826,10 @@ class _UserProfileParamsInheritedWidget extends InheritedWidget {
   final int id;
 
   static _UserProfileParamsInheritedWidget? maybeOf(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<
-        _UserProfileParamsInheritedWidget>();
+    return context
+        .dependOnInheritedWidgetOfExactType<
+          _UserProfileParamsInheritedWidget
+        >();
   }
 
   static _UserProfileParamsInheritedWidget of(BuildContext context) {
@@ -925,9 +940,10 @@ class UserProfileWidget extends ConsumerWidget {
     BuildContext context,
     UserProfileProxyWidgetRef ref,
     UserProfileState state,
-  ) builder;
+  )
+  builder;
   final void Function(UserProfileState? previous, UserProfileState next)?
-      onStateChanged;
+  onStateChanged;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -975,7 +991,8 @@ class UserProfileSelect<Selected> extends ConsumerWidget {
     BuildContext context,
     UserProfileProxyWidgetRef ref,
     Selected value,
-  ) builder;
+  )
+  builder;
   final void Function(Selected? previous, Selected next)? onStateChanged;
 
   @override
@@ -1021,7 +1038,8 @@ class UserProfileParamsBuilder extends ConsumerWidget {
     BuildContext context,
     UserProfileProxyWidgetRef ref,
     int params,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1067,7 +1085,7 @@ class UserProfileNameField extends ConsumerWidget {
   final TextEditingController? controller;
   final Duration? debounceDuration;
   final Widget Function(BuildContext context, UserProfileNameProxyWidgetRef ref)
-      builder;
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1134,7 +1152,8 @@ class UserProfileEmailField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     UserProfileEmailProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1199,7 +1218,7 @@ class UserProfileAgeField extends ConsumerWidget {
   final TextEditingController? controller;
   final Duration? debounceDuration;
   final Widget Function(BuildContext context, UserProfileAgeProxyWidgetRef ref)
-      builder;
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1264,7 +1283,7 @@ class UserProfileBioField extends ConsumerWidget {
   final TextEditingController? controller;
   final Duration? debounceDuration;
   final Widget Function(BuildContext context, UserProfileBioProxyWidgetRef ref)
-      builder;
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1328,7 +1347,7 @@ class UserProfileBio2Field extends ConsumerWidget {
   final TextEditingController? controller;
   final Duration? debounceDuration;
   final Widget Function(BuildContext context, UserProfileBio2ProxyWidgetRef ref)
-      builder;
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1398,8 +1417,10 @@ class _SecondUserProfileParamsInheritedWidget extends InheritedWidget {
   static _SecondUserProfileParamsInheritedWidget? maybeOf(
     BuildContext context,
   ) {
-    return context.dependOnInheritedWidgetOfExactType<
-        _SecondUserProfileParamsInheritedWidget>();
+    return context
+        .dependOnInheritedWidgetOfExactType<
+          _SecondUserProfileParamsInheritedWidget
+        >();
   }
 
   static _SecondUserProfileParamsInheritedWidget of(BuildContext context) {
@@ -1471,9 +1492,10 @@ class SecondUserProfileWidget extends ConsumerWidget {
     BuildContext context,
     SecondUserProfileProxyWidgetRef ref,
     UserProfileState state,
-  ) builder;
+  )
+  builder;
   final void Function(UserProfileState? previous, UserProfileState next)?
-      onStateChanged;
+  onStateChanged;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1511,7 +1533,8 @@ class SecondUserProfileSelect<Selected> extends ConsumerWidget {
     BuildContext context,
     SecondUserProfileProxyWidgetRef ref,
     Selected value,
-  ) builder;
+  )
+  builder;
   final void Function(Selected? previous, Selected next)? onStateChanged;
 
   @override
@@ -1564,7 +1587,8 @@ class SecondUserProfileNameField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     SecondUserProfileNameProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1623,7 +1647,8 @@ class SecondUserProfileEmailField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     SecondUserProfileEmailProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1682,7 +1707,8 @@ class SecondUserProfileAgeField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     SecondUserProfileAgeProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1741,7 +1767,8 @@ class SecondUserProfileBioField extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     SecondUserProfileBioProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -1799,7 +1826,8 @@ class SecondUserProfileBio2Field extends ConsumerWidget {
   final Widget Function(
     BuildContext context,
     SecondUserProfileBio2ProxyWidgetRef ref,
-  ) builder;
+  )
+  builder;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {